Remove shared lib, make single binary
Remove the build of a shared lib. Produce a single monolith
binary. This simplifies a lot the build chain.

The shared library was here to make it easier to build/compile the
tests, the test suite should not force the main project to adapt its
build chain, the test suit adapts to it.

We don't need the script AppRun anymore.